HUGE commit containing a lot of messy changes, including:
- some preliminary work for better saving of camera details (phi/chi/r)
- a hack to adjust player positions to the location a bit before the
  impact point
- changing internal texture format to RGBA for the skybox
- loading of PNG textures through UIKit on iOS
- switching to SDL2
- more OpenGL ES 2 stuff, e.g. for explosion FX, glow etc.

+++ GLtron/nebu/include/video/nebu_mesh.h
@@ -33,7 +33,7 @@ typedef struct {
 typedef struct {
 	int nPrimitivesPerIndex;
 	int nPrimitives;
-	int *pIndices;
+	unsigned short *pIndices;
 } nebu_Mesh_IB;
 
 typedef struct {
